Refactor tests into domain-specific files for clarity
Split ContractEventTests.cs into dedicated test files for merchants, statements, operators, file import logs, and estates, improving organization and maintainability. Updated SaveChangesAsync to SaveChangesWithDuplicateHandling for reconciliation events. Added ReconciliationHasStartedEvent to test data. No changes to test logic or coverage.
