+++ TransactionProcessor.Repository/ITransactionProcessorReadModelRepository.cs
@@ -1353,7 +1353,7 @@ namespace TransactionProcessor.Repository {
 
             await context.Reconciliations.AddAsync(reconciliation, cancellationToken);
 
-            return await context.SaveChangesAsync(cancellationToken);
+            return await context.SaveChangesWithDuplicateHandling(cancellationToken);
         }
 
         public async Task<Result> StartTransaction(TransactionDomainEvents.TransactionHasStartedEvent domainEvent,
